# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sunrise Staff Housing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List rooms
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: floorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomBody"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # EMPLOYEES
...
components:
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        buildingId:
          type: integer
        floorId:
          type: integer
        roomNumber:
          type: string
        roomType:
          type: string
        capacity:
          type: integer
        currentOccupancy:
          type: integer
        status:
          type: string
        gender:
          type: ["string", "null"]
      required:
        - id
        - propertyId
        - buildingId
        - floorId
        - roomNumber
        - roomType
        - capacity
        - currentOccupancy
        - status
    CreateRoomBody:
      type: object
      properties:
        propertyId:
          type: integer
        buildingId:
          type: integer
        floorId:
          type: integer
        roomNumber:
          type: string
        roomType:
          type: string
        capacity:
          type: integer
        status:
          type: string
        gender:
          type: ["string", "null"]
      required:
        - propertyId
        - buildingId
        - floorId
        - roomNumber
        - roomType
        - capacity
    UpdateRoomBody:
      type: object
      properties:
        roomNumber:
          type: string
        roomType:
          type: string
        capacity:
          type: integer
        status:
          type: string
        gender:
          type: ["string", "null"]
...